create functionality using composition API.

+++ resources/js/components/company/CompanyIndex.vue
@@ -2,9 +2,11 @@
     <div class="overflow-hidden overflow-x-auto min-w-full align-middle sm:rounded-md">
         <div class="flex place-content-end mb-4">
             <div class="px-4 py-2 text-white bg-indigo-600 hover:bg-indigo-700 cursor-pointer">
-<!--                <router-link :to="{ name: 'companies.create' }" class="text-sm font-medium">Create company</router-link>-->
+                <router-link :to="{ name: 'companies.create' }" class="text-sm font-medium">Create company</router-link>
             </div>
         </div>
+
+        <router-link :to="{ name: 'companies.create' }" class="text-sm font-medium">Create company</router-link>
 
         <table class="min-w-full border divide-y divide-gray-200">
             <thead>
@@ -78,6 +80,10 @@ export default {
         onMounted(getCompanies)
 
         const deleteCompany = async (id) => {
+
+            if(!window.confirm('Are you sure?')) {
+                return;
+            }
             await destroyCompany(id);
             await getCompanies();
         }
